requirements.txt
fastapi
requests
uvicorn[standard]
python-multipart
python-dotenv
pydantic
openai
websockets
starlette
Pillow
pyautogui
pynput
pyaudio
sounddevice
numpy
whisper
aiofiles
watchdog
httpx
